moved protoserializable code into protos package
remove various compile warnings in package.
