+++ Core/src/main/java/org/tribuo/hash/HashedFeatureMap.java
@@ -22,8 +22,6 @@ import java.util.TreeMap;
 import org.tribuo.FeatureMap;
 import org.tribuo.ImmutableFeatureMap;
 import org.tribuo.Model;
-import org.tribuo.ProtoSerializableClass;
-import org.tribuo.ProtoSerializableField;
 import org.tribuo.VariableIDInfo;
 import org.tribuo.VariableInfo;
 import org.tribuo.protos.core.FeatureDomainProto;
@@ -31,7 +29,9 @@ import org.tribuo.protos.core.HashedFeatureMapProto;
 import org.tribuo.protos.core.HasherProto;
 import org.tribuo.protos.core.ImmutableFeatureMapProto;
 import org.tribuo.protos.core.VariableInfoProto;
-import org.tribuo.util.ProtoUtil;
+import org.tribuo.protos.ProtoSerializableClass;
+import org.tribuo.protos.ProtoSerializableField;
+import org.tribuo.protos.ProtoUtil;
 
 import com.google.protobuf.Any;
 import com.google.protobuf.InvalidProtocolBufferException;
