+++ Core/src/main/java/org/tribuo/hash/MessageDigestHasher.java
@@ -25,8 +25,8 @@ import com.oracle.labs.mlrg.olcut.provenance.ObjectProvenance;
 import com.oracle.labs.mlrg.olcut.provenance.Provenance;
 import com.oracle.labs.mlrg.olcut.provenance.primitives.StringProvenance;
 
-import org.tribuo.ProtoSerializableClass;
-import org.tribuo.ProtoSerializableField;
+import org.tribuo.protos.ProtoSerializableClass;
+import org.tribuo.protos.ProtoSerializableField;
 import org.tribuo.protos.core.HasherProto;
 import org.tribuo.protos.core.MessageDigestHasherProto;
 import org.tribuo.protos.core.ModHashCodeHasherProto;
