+++ Core/src/main/java/org/tribuo/hash/ModHashCodeHasher.java
@@ -22,8 +22,8 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
 
-import org.tribuo.ProtoSerializableClass;
-import org.tribuo.ProtoSerializableField;
+import org.tribuo.protos.ProtoSerializableClass;
+import org.tribuo.protos.ProtoSerializableField;
 import org.tribuo.protos.core.HasherProto;
 import org.tribuo.protos.core.ModHashCodeHasherProto;
