+++ Core/src/main/java/org/tribuo/protos/ProtoSerializableKeysValuesField.java
@@ -1,4 +1,4 @@
-package org.tribuo;
+package org.tribuo.protos;
 
 import static java.lang.annotation.ElementType.FIELD;
